+++ xoto3/dynamodb/constants.py
@@ -1,3 +1,7 @@
 DYNAMO_MAX_RANGE_KEY_LEN = 1024
 # bytes in UTF-8 encoding.
 # https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Limits.html#limits-partition-sort-keys
+
+
+DEFAULT_ITEM_NAME = "Item"
+# used throughout xoto3